How do I error-driven-development my way through finding out which custom routes I need?
1.  Do rails generate scaffold on something.  it will make you a resource.
2.  Comment out the resource.
3. Try to visit the pages for that resource.  They will give you errors
4. Write a custom route based on the error.

Pro tip- if you want to know what routes rails is expecting, put the resource back in the routes file and run rake routes. It will show you all of the routes and the conventional names for those resrouces. You can then take those routes, the names, and the controller functions it is expecting and stick them in a custom route described here: http://guides.rubyonrails.org/routing.html#naming-routes